Separate enums and types
