+++ src/miscs/constants.ts
@@ -1,0 +1,5 @@
+export enum ContentEnum {
+  index = '/index',
+  game = '/game',
+  ranking = '/ranking',
+}
