+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react';
 import { useRecoilValue } from 'recoil';
 import { contentState } from '@/store/atoms';
 import { LoadingContent, IndexContent, GameContent, RankingContent } from '@/components/contents';
-import { ContentEnum } from '@/types';
+import { ContentEnum } from '@/miscs/constants';
 
 export default function Home() {
   const content = useRecoilValue(contentState);
